media-gfx/renderman: Added license template
This is not a valid license, but instead just a template of the
pixar.license file, which you get after installing renderman, stripped
off of any user data and an empty feature list. It's only there to
have a valid license file linked in the ebuild, to be able to emerge
the package.